Refactor login page name
